Only remove a stream from the routing table if it dangles. That is, there are no listeners nor source stream.
This fixes a crash that occurs after this sequence:
  1.  Connect publisher
  2. Connect client
  3. Kill publisher
  4. Reconnect publisher
  5. Kill client

This is because the listening client tries to remove itself from
a publishing list that it does not belong to.

Since streams are now persistent between publishers, a list of
clients can now continue playback even when a publisher
disconnects and reconnects.